--- src/addqueue.html
@@ -1,9 +1,0 @@
-<form class="search-form" id="search-form">
-    
-    <button type="submit" class="queue">Queue</button>
-</form>
-<div class="container">
-
-</div>
-
-<script src="./addqueue.js" type="module"></script>
