+++ src/index.html
@@ -24,5 +24,6 @@
     <include src="./partials/header_home.html"></include>
     <include src="./partials/modal.html"></include>
     <include src="./partials/pagination.html"></include>
+    <include src="./partials/addqueue.html"></include>
   </body>
 </html>
