+++ src/index.js
@@ -1,4 +1,4 @@
 
 import './js/services/services';
 import './js/modal';
-
+import './js/addqueue';
