+++ src/partials/addqueue.html
@@ -1,0 +1,9 @@
+<!-- <form class="search-form" id="search-form">
+    
+    <button type="submit" class="queue">Queue</button>
+</form> -->
+<div class="container">
+
+</div>
+
+<script src="./js/addqueue.js" type="module"></script>
