+++ src/sass/_addqueue.scss
@@ -1,46 +1,13 @@
-.container {
-    margin-left: auto;
-    margin-right: auto;
-
-    padding-left: 32px;
-    padding-right: 32px;
-
-    @media screen and (min-width: $mob) {
-        padding-left: 20px;
-        padding-right: 20px;
-        }
-
-    // outline: 2px solid red;
-    // outline-offset: -2px;
-
-    // mobile screen
-
-    @media screen and (min-width: $mob) {
-        width: $mob;
-    }
-
-    // tablet screen
-    @media screen and (min-width: $tab) {
-        width: $tab;
-    }
-
-    // desktop screen
-    @media screen and (min-width: $desk) {
-        width: $desk;
-    }
-
-}
-
 .gallery {
     display: flex;
     flex-wrap: wrap;
-    @media screen and (min-width: $tab) and (max-width: ($desk - 0.02)) {
+    @media screen and (min-width: 768px) and (max-width: (1280px - 0.02)) {
         gap: 32px;
     }
-    @media screen and (min-width: $desk) {
+    @media screen and (min-width: 1280px) {
         gap: 16px;
     }
-    @media screen and (min-width: $mob) and (max-width: ($tab - 0.02)) {
+    @media screen and (min-width: 320px) and (max-width: (768px - 0.02)) {
         display: block;
     }
 }
@@ -51,15 +18,15 @@
     font-style: normal;
     font-weight: 500;
     font-size: 12px;
-    @media screen and (min-width: $mob) and (max-width: ($tab - 0.02)) {
+    @media screen and (min-width: 320px) and (max-width: (768px - 0.02)) {
         line-height: 14px;
         margin-bottom: 20px;
     }
-    @media screen and (min-width: $tab) and (max-width: ($desk - 0.02)) {
+    @media screen and (min-width: 768px) and (max-width: (1280px - 0.02)) {
         line-height: 16px;
         margin-bottom: 32px;
     }
-    @media screen and (min-width: $desk) {
+    @media screen and (min-width: 1280px) {
         font-size: 20px;
         line-height: 24px;
         /* or 120% */
@@ -70,9 +37,9 @@
 }
 
 .info__item {
-    color: $black-color;
+    color: #000000;
 }
 
 .info__red {
-    color: $red-color;
+    color: #b92f2c;
 }
