+++ src/sass/index.scss
@@ -6,3 +6,5 @@
 @import './header';
 @import './modal';
 @import './pagination';
+@import './addqueue';
+
