+++ src/sass/utils/_reset.scss
@@ -41,18 +41,18 @@ img {
   margin: 0 auto;
   padding-left: 20px;
   padding-right: 20px;
-  @media screen and (min-width: 480px) {
-    width: 480px;
+  @media screen and (min-width: $mob) {
+    width: $mob;
   }
 
-  @media screen and (min-width: 768px) {
-    width: 768px;
+  @media screen and (min-width: $tab) {
+    width: $tab;
     padding-left: 32px;
     padding-right: 32px;
   }
 
-  @media screen and (min-width: 1200px) {
-    width: 1200px;
+  @media screen and (min-width: $desk) {
+    width: $desk;
     padding-left: 32px;
     padding-right: 32px;
   }
